Fix combo box showing only None when options change by calling filter in ngOnChanges
The custom combo box implements a lazy-loading feature, meaning the list of options appearing in the *ngFor loop in the template for the mat-options does not match the list of options bound to the component. If the options input array changes when the mat-select is open, therefore, the new list of options will not appear unless the mat-select is first closed and re-opened. This is fixed by checking if the mat-select is open in the ngOnChanges function and re-calling the appropriate filter function.